refactor DefaultPrepareForPreviewUseCase using converters
